fix: match OLA padding dtype to input chunk in eq_processor (#2883)
np.zeros() defaulted to float64, silently upcasting the final chunk
when input is float32. Now uses chunk.dtype to preserve the input
precision. The other np.zeros (gain curve at line 213) correctly
uses float64 for intermediate dB computation.

+++ auralis/core/processing/eq_processor.py
@@ -178,7 +178,7 @@ class EQProcessor:
 
             # Pad chunk if necessary
             if len(chunk) < chunk_size:
-                padded_chunk = np.zeros((chunk_size,) + chunk.shape[1:])
+                padded_chunk = np.zeros((chunk_size,) + chunk.shape[1:], dtype=chunk.dtype)
                 padded_chunk[:len(chunk)] = chunk
                 chunk = padded_chunk
